refactor: extract generic ResumeBox<Decision> shared by trust coordinators
TLSTrustCoordinator and HostKeyMismatchCoordinator carried byte-identical private ResumeBox copies (the second is where a dismissible-sheet regression entered). Lift the single-resume continuation box into one generic ResumeBox<Decision> in Pulse/Concurrency/, consumed by both. Behaviour-preserving: the existing coordinator suites pass unchanged.
